+++ src/main/kotlin/codegen/FlutterXPluginInfo.kt
@@ -1,12 +1,12 @@
 package codegen
 
 object FlutterXPluginInfo {
-    const val Version: String = "5.2.3.243"
+    const val Version: String = "5.2.3.242"
     const val Changelog: String = """
-<h2>5.2.3 - 2025-02-06</h2>
+<h2>5.2.3 - 2025-02-13</h2>
 
-<ul><li>Optimize known issues</li></ul>
+<ul><li>Replaced flutter version update record link</li><li>Fix the wrong command to open an ios project</li><li>Optimize known issues</li></ul>
 """
 
-    const val BuildDate: String = "2025-02-06"
+    const val BuildDate: String = "2025-02-13"
 }